// src/app/models/Cita/triage.model.ts
export class triage {
    id: number=0;
    idCita: number;
    signosSintomas: string;
    presionArterial: string;
    frecuenciaCardiaca: number;
    temperatura: number;
    saturacionOxigeno: number;
    peso: number;
    talla: number;
    antecedentesMedicos: string;
    diagnosticoProvisional: string;
    tratamientoRecomendado: string;
    idSeguro: number;
    observaciones: string;
}